Fix conflicts arising from merge with master
convert int fields for class participation score to double

+++ src/main/java/seedu/taskmaster/model/ModelManager.java
@@ -304,7 +304,7 @@ public class ModelManager implements Model {
 
     @Override
     public void showLowestScoringStudents() {
-        int lowestScore = taskmaster.getLowestScore();
+        double lowestScore = taskmaster.getLowestScore();
         filteredStudentRecords.setPredicate(new ScoreEqualsPredicate(lowestScore));
     }
